Clean up build warning.

+++ src/Math/SSEMath.cpp
@@ -14,7 +14,7 @@ void *AlignedMalloc(size_t size, size_t alignment)
 	if (!ptr)
 		return 0;
 	++ptr; // Must make room for storing the offset info.
-	int incr = (alignment - (ptr & (alignment-1))) & (alignment-1);
+	ptrdiff_t incr = (alignment - (ptr & (alignment-1))) & (alignment-1);
 	ptr += incr;
 	((u8*)ptr)[-1] = (u8)(incr+1);
 	return (void*)ptr;
